LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.fillStyle.parse.current.notrendered-expected.txt

2d.fillStyle.parse.current.basic
currentColor is computed from the canvas element
Actual output:
Expected output:


PASS currentColor is computed from the canvas element even when element is not rendered 
